// app/src/api/model/correlationMethod.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Economicon App API
 * OpenAPI spec version: 0.2.0
 */

export type CorrelationMethod = typeof CorrelationMethod[keyof typeof CorrelationMethod];


export const CorrelationMethod = {
  pearson: 'pearson',
  spearman: 'spearman',
  kendall: 'kendall',
} as const;