// app/src/api/model/missingHandlingMethod.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Economicon App API
 * OpenAPI spec version: 0.3.0
 */

export type MissingHandlingMethod = typeof MissingHandlingMethod[keyof typeof MissingHandlingMethod];


export const MissingHandlingMethod = {
  pairwise: 'pairwise',
  listwise: 'listwise',
} as const;